[TEST] Extract shared context to decidim-admin.
