+++ decidim-admin/lib/decidim/admin/test.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+require "decidim/admin/test/with_admin_filters_context"
 require "decidim/admin/test/manage_attachments_examples"
 require "decidim/admin/test/manage_attachment_collections_examples"
 require "decidim/admin/test/manage_component_permissions_examples"
