+++ decidim-proposals/spec/system/admin_filters_proposals_spec.rb
@@ -2,27 +2,6 @@
 
 require "spec_helper"
 
-shared_context "with admin filters" do
-  # @param label: The aria-label in the LI tags to be hovered of the filter.
-  # @param text_to_click: The text to click in the opened submenu of the filter.
-  def visit_filtered_results(label, text_to_click)
-    within ".filters__section" do
-      find("li", text: "FILTER").hover
-      find("li[aria-label=#{label}]").hover
-      within("li[aria-label=#{label}") do
-        find("a", text: text_to_click).click
-      end
-    end
-  end
-
-  def visit_filtered_results_by_search(text)
-    within ".filters__section" do
-      fill_in :q_id_or_title_cont, with: text
-      find("*[type=submit]").click
-    end
-  end
-end
-
 describe "Admin filters proposals", type: :system do
   let(:manifest_name) { "proposals" }
   let!(:proposal) { create :proposal, component: current_component }
